Backend/load-test/requirements.txt
Flask==2.3.2
locust==2.15.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
